# cmd/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Documentation for Login Gateway Service",
        "title": "Login Gateway API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9001",
    "basePath": "/",
    "paths": {
...
        "/mintegra/notification/post": {
            "post": {
                "description": "Send POST Notification to WebSocket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Websocket"
                ],
                "summary": "POST Notification",
                "parameters": [
                    {
                        "description": "Data POST Notification",
                        "name": "notification",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_notification.NotificationMessage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "POST Notification sent!",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid JSON format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to send notification",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/mintegra/notification/pub": {
            "get": {
                "description": "Get Publish Notification",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Websocket"
                ],
                "summary": "Publish Notification WebSocket",
                "responses": {
                    "101": {
                        "description": "WebSocket connection established",
                        "schema": {
                            "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_notification.NotificationMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_response.MainResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_response.MainResponse"
                        }
                    }
                }
            }
        },
        "/mintegra/notification/sos": {
            "post": {
                "description": "Send SOS Notification to WebSocket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Websocket"
                ],
                "summary": "SOS Notification",
                "parameters": [
                    {
                        "description": "Data SOS Notification",
                        "name": "notification",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_notification.NotificationMessage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SOS Notification sent!",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid JSON format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to send notification",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/mintegra/notification/sub": {
            "get": {
                "description": "Get Subscribe Notification",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Websocket"
                ],
                "summary": "Subscribe to Notification WebSocket",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel format. Contoh nilai yang valid:\n1. customer_no:site:master-data\n2. customer_no:site:opr-dispatch\n3. customer_no:site:opr-report\n4. customer_no:site:opr-fleet\n5. customer_no:site:plant-oem\n6. customer_no:site:all",
                        "name": "channel",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "WebSocket connection established",
                        "schema": {
                            "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_notification.NotificationMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_response.MainResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_response.MainResponse"
                        }
                    }
                }
            }
        },
        "/mintegra/notification/test": {
            "post": {
                "description": "Test Notification to WebSocket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Websocket"
                ],
                "summary": "Test Notification",
                "parameters": [
                    {
                        "description": "Data Test Notification",
                        "name": "notification",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_notification.NotificationMessage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification sent!",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid JSON format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to send notification",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_royanqodri_Login-Gateway-API_model_notification.NotificationMessage": {
            "type": "object",
            "properties": {
                "activity": {
                    "type": "string"
                },
                "altitude": {
                    "type": "number"
                },
                "bearing": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "channel": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "customer_no": {
                    "type": "string"
                },
                "date_log": {
                    "type": "string"
                },
                "equipment_model": {
                    "type": "string"
                },
                "equipment_no": {
                    "type": "string"
                },
                "equipment_type": {
                    "type": "string"
                },
                "event": {
                    "type": "string"
                },
                "fleet": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "shift": {
                    "type": "string"
                },
                "shift_sequence": {
                    "type": "integer"
                },
                "site": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "status_activity": {
                    "type": "string"
                },
                "time_log": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "github_com_royanqodri_Login-Gateway-API_model_response.MainResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "data_header": {},
                "status_response": {
                    "$ref": "#/definitions/github_com_royanqodri_Login-Gateway-API_model_response.StatusResponse"
                },
                "total_data": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_royanqodri_Login-Gateway-API_model_response.StatusResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                },
                "status_text": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}